Update on 2017/08/23 1. Fixed the issue that the game would be out of control after some instructions in the guide mode game. 2. Modified some codes so that the projest is now compatible with cocos2d-3.15.1. 3. Replaced some big size audios with smaller ones.

+++ Classes/InputManager.cpp
@@ -51,13 +51,13 @@ void InputManager::onKeyReleased(EventKeyboard::KeyCode keyCode, Event * event){
 
 void InputManager::onMouseUp(Event *event){
 	_isKeyPressed[(int)InputManager::KeyCode::KEY_LEFT_MOUSE_BUTTON \
-		+ ((EventMouse *)event)->getMouseButton()] = false;
+		+ (int)((EventMouse *)event)->getMouseButton()] = false;
 }
 
 void InputManager::onMouseDown(Event *event){
 	
 	_isKeyPressed[(int)InputManager::KeyCode::KEY_LEFT_MOUSE_BUTTON \
-		+ ((EventMouse *)event)->getMouseButton()] = true;
+		+ (int)((EventMouse *)event)->getMouseButton()] = true;
 }
 
 void InputManager::onMouseMove(Event *event){
